refactor: move documentation components in api package
