+++ gravitee-apim-console-webui/src/management/api/apis.module.ts
@@ -44,6 +44,7 @@ import { ApiFailoverModule } from './failover/api-failover.module';
 import { ApiHealthCheckModule } from './health-check/api-health-check.module';
 import { ApiPlansModule } from './plans/api-plans.module';
 import { ApiSubscriptionsModule } from './subscriptions/api-subscriptions.module';
+import { ApiDocumentationModule } from './documentation/api-documentation.module';
 
 import { SpecificJsonSchemaTypeModule } from '../../shared/components/specific-json-schema-type/specific-json-schema-type.module';
 import { DocumentationModule } from '../../components/documentation/documentation.module';
@@ -62,6 +63,7 @@ import { AlertsModule } from '../../components/alerts/alerts.module';
     ApiCreationGetStartedModule,
     ApiCreationV2Module,
     ApiCreationV4Module,
+    ApiDocumentationModule,
     ApiDocumentationV4Module,
     ApiEndpointsModule,
     ApiEndpointsV4Module,
