+++ gravitee-apim-console-webui/src/management/api/documentation/api-documentation.module.ts
@@ -18,11 +18,11 @@ import { CommonModule } from '@angular/common';
 
 import { ApiPortalDocumentationMetadataComponent } from './metadata/api-portal-documentation-metadata.component';
 
-import { GioMetadataModule } from '../../../../components/gio-metadata/gio-metadata.module';
+import { GioMetadataModule } from '../../../components/gio-metadata/gio-metadata.module';
 
 @NgModule({
   declarations: [ApiPortalDocumentationMetadataComponent],
   exports: [ApiPortalDocumentationMetadataComponent],
   imports: [CommonModule, GioMetadataModule],
 })
-export class ApiGeneralDocumentationModule {}
+export class ApiDocumentationModule {}
