+++ gravitee-apim-console-webui/src/management/api/documentation/metadata/api-portal-documentation-metadata.component.html
