+++ gravitee-apim-console-webui/src/management/api/documentation/metadata/api-portal-documentation-metadata.component.spec.ts
@@ -25,13 +25,13 @@ import { ActivatedRoute } from '@angular/router';
 
 import { ApiPortalDocumentationMetadataComponent } from './api-portal-documentation-metadata.component';
 
-import { CONSTANTS_TESTING, GioHttpTestingModule } from '../../../../../shared/testing';
-import { Metadata } from '../../../../../entities/metadata/metadata';
-import { fakeMetadata } from '../../../../../entities/metadata/metadata.fixture';
-import { GioMetadataHarness } from '../../../../../components/gio-metadata/gio-metadata.harness';
-import { ApiGeneralDocumentationModule } from '../api-general-documentation.module';
-import { GioMetadataDialogHarness } from '../../../../../components/gio-metadata/dialog/gio-metadata-dialog.harness';
-import { GioTestingPermissionProvider } from '../../../../../shared/components/gio-permission/gio-permission.service';
+import { CONSTANTS_TESTING, GioHttpTestingModule } from '../../../../shared/testing';
+import { Metadata } from '../../../../entities/metadata/metadata';
+import { fakeMetadata } from '../../../../entities/metadata/metadata.fixture';
+import { GioMetadataHarness } from '../../../../components/gio-metadata/gio-metadata.harness';
+import { ApiDocumentationModule } from '../api-documentation.module';
+import { GioMetadataDialogHarness } from '../../../../components/gio-metadata/dialog/gio-metadata-dialog.harness';
+import { GioTestingPermissionProvider } from '../../../../shared/components/gio-permission/gio-permission.service';
 
 describe('ApiPortalDocumentationMetadataComponent', () => {
   let fixture: ComponentFixture<ApiPortalDocumentationMetadataComponent>;
@@ -47,7 +47,7 @@ describe('ApiPortalDocumentationMetadataComponent', () => {
         { provide: ActivatedRoute, useValue: { snapshot: { params: { apiId: API_ID } } } },
         { provide: GioTestingPermissionProvider, useValue: ['api-metadata-u', 'api-metadata-d', 'api-metadata-c'] },
       ],
-      imports: [NoopAnimationsModule, GioHttpTestingModule, ApiGeneralDocumentationModule, MatIconTestingModule],
+      imports: [NoopAnimationsModule, GioHttpTestingModule, ApiDocumentationModule, MatIconTestingModule],
     })
       .overrideProvider(InteractivityChecker, {
         useValue: {
